chore: wagmi examples
