+++ examples/ethereum-provider-wagmi/README.md
