+++ examples/ethereum-provider-wagmi/index.html
