+++ examples/ethereum-provider-wagmi/public/appkit-logo.png
