+++ examples/ethereum-provider-wagmi/public/favicon-dark.png
