+++ examples/ethereum-provider-wagmi/public/favicon.png
