+++ examples/ethereum-provider-wagmi/src/app.tsx
@@ -4,7 +4,7 @@ import { Connectors } from './components/Connectors'
 export default function App() {
   return (
     <main>
-      <h1>WCM Wagmi</h1>
+      <h1>Ethereum Provider Wagmi Example</h1>
       <Connectors />
       <hr />
       <Account />
