+++ examples/ethereum-provider-wagmi/src/env.d.ts
