+++ examples/ethereum-provider-wagmi/tsconfig.json
