+++ examples/ethereum-provider-wagmi/vite.config.ts
