+++ examples/react-wagmi/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "1.2.0",
   "scripts": {
-    "dev": "vite --port 3002",
+    "dev": "vite",
     "build": "vite build"
   },
   "dependencies": {
