--- examples/wcm-wagmi/src/main.tsx
@@ -1,34 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { WagmiProvider, createConfig, http } from 'wagmi'
-import { mainnet, sepolia } from 'wagmi/chains'
-
-import App from './app.jsx'
-import './assets/main.css'
-import {
-  customWalletConnectConnector,
-  defaultWalletConnectConnector
-} from './connectors/wallet-connect-connector.js'
-
-const config = createConfig({
-  chains: [mainnet, sepolia],
-  transports: {
-    [mainnet.id]: http(),
-    [sepolia.id]: http()
-  },
-  connectors: [defaultWalletConnectConnector, customWalletConnectConnector]
-})
-
-const queryClient = new QueryClient()
-
-ReactDOM.createRoot(document.getElementById('app')!).render(
-  <React.StrictMode>
-    <WagmiProvider config={config}>
-      <QueryClientProvider client={queryClient}>
-        <App />
-      </QueryClientProvider>
-    </WagmiProvider>
-  </React.StrictMode>
-)
